+++ com.unity.render-pipelines.universal/Shaders/2D/Light2D-Shape-Volumetric.shader
@@ -34,23 +34,23 @@ Shader "Hidden/Light2D-Shape-Volumetric"
             struct Varyings
             {
                 float4  positionCS  : SV_POSITION;
-                half4   color       : COLOR;
-                half2   uv          : TEXCOORD0;
+                float4  color       : COLOR;
+                float2  uv          : TEXCOORD0;
 
                 SHADOW_COORDS(TEXCOORD1)
             };
 
-            half4 _LightColor;
-            half  _FalloffDistance;
-            half4 _FalloffOffset;
-            half  _VolumeOpacity;
-            half  _InverseHDREmulationScale;
+            float4 _LightColor;
+            float  _FalloffDistance;
+            float4 _FalloffOffset;
+            float  _VolumeOpacity;
+            float  _InverseHDREmulationScale;
 
 #ifdef SPRITE_LIGHT
             TEXTURE2D(_CookieTex);			// This can either be a sprite texture uv or a falloff texture
             SAMPLER(sampler_CookieTex);
 #else
-            uniform half  _FalloffIntensity;
+            uniform float  _FalloffIntensity;
             TEXTURE2D(_FalloffLookup);
             SAMPLER(sampler_FalloffLookup);
 #endif
